fixed build warning messages.  Since, in this context, the output of ms_int32 is an int.
git-svn-id: http://svn.osgeo.org/mapserver/trunk@6209 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ shptreevis.c
@@ -158,7 +158,7 @@ int main( int argc, char ** argv )
 #endif
 
     printf ("This %s %s index supports a shapefile with %d shapes, %d depth \n",
-	(qix->version ? "new": "old"), (qix->LSB_order? "LSB": "MSB"), qix->nShapes, qix->nDepth);
+	(qix->version ? "new": "old"), (qix->LSB_order? "LSB": "MSB"), (int) qix->nShapes, (int) qix->nDepth);
 
 
 /* -------------------------------------------------------------------- */
